[wip] Feature/history 3661

+++ src/components/history/messages/RegularMessage.js
@@ -7,7 +7,7 @@ import copy from 'copy-to-clipboard'
 import Avatar from '../../avatar/Avatar'
 import Grapedown from '../../grapedown/Grapedown'
 import Header from '../../message-parts/Header'
-import {OwnBubble, MateBubble} from './Bubble'
+import {OwnBubble, MateBubble, EditModeBubble} from './Bubble'
 import Menu from '../../message-parts/Menu'
 import ImageAttachment from '../../message-parts/attachments/ImageAttachment'
 import LinkAttachment from '../../message-parts/attachments/LinkAttachment'
@@ -47,6 +47,7 @@ export default class RegularMessage extends Component {
     isPending: PropTypes.bool.isRequired,
     isUnsent: PropTypes.bool.isRequired,
     isOwn: PropTypes.bool.isRequired,
+    editMode: PropTypes.bool.isRequired,
     link: PropTypes.string.isRequired,
     onEdit: PropTypes.func.isRequired,
     onRemove: PropTypes.func.isRequired,
@@ -64,6 +65,7 @@ export default class RegularMessage extends Component {
     isUnsent: false,
     hasBubbleArrow: true,
     isOwn: false,
+    editMode: false,
     attachments: [],
     children: '',
     onEdit: noop,
@@ -88,7 +90,9 @@ export default class RegularMessage extends Component {
         copy(this.props.link)
         break
       case 'remove':
-        this.props.onRemove()
+        if (confirm('Delete the selected Message?')) { // eslint-disable-line no-alert
+          this.props.onRemove()
+        }
         break
       case 'edit':
         this.props.onEdit()
@@ -98,9 +102,11 @@ export default class RegularMessage extends Component {
   }
 
   renderMenu = () => {
+    const {isOwn, attachments, sheet, isPending, isUnsent} = this.props
+
+    if (isPending || isUnsent) return null
     if (!this.state.isMenuOpened) return null
 
-    const {isOwn, attachments, sheet} = this.props
     let items
 
     if (isOwn) {
@@ -132,10 +138,15 @@ export default class RegularMessage extends Component {
   render() {
     const {
       sheet, author, time, userTime, avatar, children, hasBubbleArrow,
-      isPending, isOwn, isUnsent, onResend, attachments
+      isPending, isOwn, isUnsent, editMode, onResend, attachments
     } = this.props
     const {classes} = sheet
-    const ThemedBubble = isOwn ? OwnBubble : MateBubble
+    let Bubble
+    if (editMode) {
+      Bubble = EditModeBubble
+    } else {
+      Bubble = isOwn ? OwnBubble : MateBubble
+    }
     return (
       <div className={classes.message}>
         {author &&
@@ -150,13 +161,13 @@ export default class RegularMessage extends Component {
           onMouseEnter={this.onMouseEnter}
           onMouseLeave={this.onMouseLeave}>
           {avatar && <Avatar src={avatar} className={classes.avatar} />}
-          <ThemedBubble className={classes.bubble} hasArrow={hasBubbleArrow}>
-            <div className={isPending ? classes.pending : ''}>
+          <Bubble className={classes.bubble} hasArrow={hasBubbleArrow}>
+            <div className={`${classes.content} ${isPending || isUnsent ? classes.pending : ''}`}>
               <Grapedown text={children} />
               {attachments.map(this.renderAttachment)}
             </div>
             {this.renderMenu()}
-          </ThemedBubble>
+          </Bubble>
         </div>
         {isUnsent && <Unsent theme={{classes}} onResend={onResend} />}
       </div>
